Create search bar in Glossary
