+++ src/output.css
@@ -31,41 +31,6 @@ body,
   width: 50%;
   font-size: 4rem;
 }
-.app__header form {
-  display: flex;
-  align-items: center;
-  width: 40%;
-}
-.app__header input,
-.app__header button {
-  height: 3rem;
-  padding: 0 0.5em;
-  border-radius: 2px;
-  border: none;
-}
-.app__header input[type=text] {
-  width: 40%;
-  margin: 0.5em;
-  font-family: "Sarala", sans-serif;
-}
-.app__header button {
-  width: 15%;
-  margin: 0;
-  padding: 0;
-  background-color: #8D99AE;
-  cursor: pointer;
-}
-.app__header label,
-.app__header button {
-  font-size: 2rem;
-  font-family: "Caudex", serif;
-  letter-spacing: 0.75px;
-  color: #EDF2F4;
-}
-.app__header button:hover {
-  background-color: #EDF2F4;
-  color: #EF233C;
-}
 
 .app__body {
   height: 85%;
@@ -169,6 +134,33 @@ main .main__loading {
   align-items: center;
   justify-content: center;
   animation: fadeCycle 1.5s infinite;
+}
+main .main__glossary {
+  width: 80%;
+}
+main .main__search {
+  display: flex;
+  align-items: center;
+  width: 100%;
+}
+main input {
+  font-family: "Sarala", sans-serif;
+  font-size: 1.5rem;
+  width: 50%;
+  height: 3.5rem;
+  margin: 0.5em;
+  padding: 0 0.5em;
+  border-radius: 2px;
+  border: 1px solid #C4CAD6;
+}
+main input:hover {
+  border: 1px solid #EF233C;
+}
+main label {
+  font-size: 2.2rem;
+  font-family: "Caudex", serif;
+  letter-spacing: 0.75px;
+  color: #2B2D42;
 }
 
 @keyframes fadeCycle {
